Add condition only_has_authors_in
"only_has_authors_in" is satisfied when all relevant commits are
authored by users in the users list or that belong to any of the
listed organizations or teams.

This differs from "only_has_contributors_in" in that it does not
take into account the committers, and allows filtering on authors
only.

Change-type: minor
Signed-off-by: Kyle Harding <kyle@balena.io>