Fix: Lookup missing verification properties on PR commit list
This was working for deployment protection events because
only the tip of the branch was considered for approval.

But the logs showed that when evaluating all of the commits
on a PR, the verification properties were undefined.

This is due to the different schema returned by the listCommits
endpoint vs the getCommit endpoint we had not accounted for.

Change-type: patch
Signed-off-by: Kyle Harding <kyle@balena.io>